requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pyreadstat>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
python-docx>=0.8.11
mammoth>=1.6.0
werkzeug>=2.3.0
